+++ test/dummy/test/test_helper.rb
@@ -4,3 +4,5 @@ ENV["RAILS_ENV"] ||= "test"
 
 require_relative "../config/environment"
 require "rails/test_help"
+
+Dir[Rails.root.join("test/support/**/*.rb")].sort.each { |path| require path }